refactor: Reorganise and remove some types
