--- src/client/types/FeedbackTargets.ts
@@ -1,3 +1,0 @@
-import { FeedbackTarget } from './FeedbackTarget'
-
-export type FeedbackTargets = FeedbackTarget[]
