--- src/client/types/InitialValues.ts
@@ -1,10 +1,0 @@
-import { Locales } from './Locales'
-
-export interface InitialValues {
-  name: Locales
-  startDate: Date
-  endDate: Date
-  studentNumbers: []
-  teachers: []
-  courses: []
-}
