--- src/client/types/OrganisationCode.ts
@@ -1,1 +1,0 @@
-export type OrganisationCode = string
